requirements.txt
python-dateutil
arrow
gevent>=1.0
IPy
python-decouple

pymongo>=2.8,<3.0
mongoengine>=0.9.0
#Flask-Mongoengine - bug tz_aware sur session
https://github.com/MongoEngine/flask-mongoengine/tarball/master
mongolock>=1.3.3

#flanker
https://github.com/srault95/flanker/tarball/light_deps

Flask-Script
Flask-Login<0.3.0
Flask-Security
Flask-Moment
Flask-BabelEx
Flask-Admin>=1.1.0

#geoip-data
https://github.com/srault95/geoip-data/tarball/master

pygeoip

https://github.com/srault95/mongo-mail-tools/tarball/master
Flask-DebugToolbar